web_server/Project/requirements.txt
flask~=1.1.2
tensorflow-gpu~=2.3.0
numpy~=1.18.5
pandas~=1.2.3
scipy~=1.4.1